Changed reward to float64.

+++ dsr/dsr/task/control/control.py
@@ -141,7 +141,7 @@ def make_control_task(function_set, name, action_spec, algorithm=None,
         """Runs n_episodes episodes and returns each episodic reward."""
 
         # Run the episodes and return the average episodic reward
-        r_episodes = np.zeros(n_episodes, dtype=np.float32) # Episodic rewards for each episode
+        r_episodes = np.zeros(n_episodes, dtype=np.float64) # Episodic rewards for each episode
         for i in range(n_episodes):
 
             # During evaluation, always use the same seeds
